Made radios more modular
